fix[verify]: cosig threshold compared in i64, not usize — close wasm32 fail-open
Adversarial-review finding (GLM 5.2 standing in for the rate-limited Codex gate,
verified against the code + Opus-reviewed SOUND): the cosig (M6) M-of-N gate
decided satisfaction with `approvals >= cosig_threshold as usize`. On the shipped
wasm32 verifier (the browser/auditor surface) `usize == u32`, so a broker-signed
`cosig_threshold` above 2^32 truncates: 2^32+1 -> 1, and a SINGLE approver
satisfies it (1 >= 1) -> a sub-threshold grant indexes and reaches ok:true /
cosig_status:"satisfied", defeating dual control. The challenge preimage uses a
full-width BE8 threshold, so the cosignatures verify; only the comparison
truncated. The 64-bit CLI/cgo build was unaffected — a platform-dependent verdict,
itself a soundness violation. My earlier Opus passes analyzed the cast only on
64-bit and missed the 32-bit case; a different reviewer (GLM) caught it.

Fix: compare in i64 — `approvals as i64 >= cosig_threshold` (approvals is a small
bounded count, so the widen is lossless on every target; cosig_threshold>=1 is
already guaranteed). Same truncation existed pre-existing at the bounded_reuse
overspend check (`use_limit as usize`) — fixed too (there it shrank a max, i.e.
fail-CLOSED, but the verdict must still be platform-independent). grep confirms no
other untrusted-i64 `as usize` truncation remains.

New regression test tier_b_cosig_threshold_above_u32_is_not_truncated
(cosig_threshold=2^32+1, 1 approver -> unsatisfied / !ok). core 153 adversarial +
workspace + wasm32 build + clippy green. (Shipped verifier/*.wasm is gitignored,
rebuilt from source at deploy, so the fix propagates to the browser verifier.)

+++ core/src/verify.rs
@@ -2477,7 +2477,12 @@ pub fn verify_bundle_with(bundle: &CanonValue, opts: &VerifyOptions) -> VerifyRe
                         exp,
                         &opts.cosig_approver_keys,
                     );
-                    let satisfied = approvals >= cosig_threshold as usize;
+                    // Compare in i64, NOT `cosig_threshold as usize`: on the wasm32 verifier `usize == u32`,
+                    // so a broker-signed `cosig_threshold` above 2^32 would truncate to a small value and let a
+                    // SUB-threshold grant read as satisfied (a fail-open on the browser verifier; the 64-bit
+                    // CLI/cgo build was unaffected). `approvals` is bounded by the cosignature count, so the
+                    // widen is lossless on every target. (`cosig_threshold >= 1` is already guaranteed above.)
+                    let satisfied = approvals as i64 >= cosig_threshold;
                     if cosig_seen.insert(gid.clone()) {
                         cosigned_grants_total += 1;
                         if satisfied {
@@ -2848,7 +2853,10 @@ pub fn verify_bundle_with(bundle: &CanonValue, opts: &VerifyOptions) -> VerifyRe
                 violation(&mut issues, format!("bounded_reuse grant '{gid}' use_sequence_number {usn} replayed across receipts — seq-replay (M1)"));
                 continue;
             }
-            if g.used >= g.use_limit as usize {
+            // Compare in i64, NOT `use_limit as usize` (same wasm32 `usize==u32` truncation concern as the
+            // cosig gate above). Here truncation would shrink the cap (fail-CLOSED, over-restrictive) rather
+            // than open, but the verdict must still be platform-independent; `g.used` is a small count.
+            if g.used as i64 >= g.use_limit {
                 bounded_reuse_overspent += 1;
                 unmatched_violation += 1;
                 violation(&mut issues, format!("bounded_reuse grant '{gid}' exercised more than use_limit {} times — overspend (M1)", g.use_limit));
